Setting up tokens for the deletion and update of posts. Adding functionality when creating posts.
- Implementation of the names of images of the random posts and deletion of those as well as the comments correpsondant to the post during its deletion

+++ public/src/Controllers/DeletePost.php
@@ -4,7 +4,10 @@ namespace Application\Controllers;
 
 use Application\Lib\Redirect;
 use Application\Lib\DatabaseConnection;
+use Application\Model\Repository\CommentRepository;
+use Application\Model\Repository\UsersRepository;
 use Application\Model\Repository\PostRepository;
+use Exception;
 
 class DeletePost
 {
@@ -12,14 +15,27 @@ class DeletePost
     {
         $connection = new DatabaseConnection();
         $postRepository = new PostRepository($connection);
+        $userRepository = new UsersRepository($connection);
+        $commentRepository = new CommentRepository($connection, $userRepository, $postRepository);
 
-        $post = $postRepository->getPosts();
         $redirection = new Redirect();
 
+        if(isset($_POST['token']) && $_POST['token'] === $_SESSION['token']) {
 
-        $postRepository->deletePost($_GET['id']);
 
-        $redirection->execute('index.php?action=posts');
+            $post = $postRepository->getPostById($_GET['id']);
+
+            $commentRepository->deleteCommentByPost($_GET['id']);
+            $postRepository->deletePost($_GET['id']);
+
+            if (file_exists($post->picture)) {
+                unlink($post->picture);
+            }
+
+            $redirection->execute('index.php?action=posts');
+        } else {
+            throw new Exception('Jeton de sécurité périmé');
+        }
     }
 }
